Update function included
Some minor fixes left. Buttons working as expected.

+++ app/src/main/java/com/ggonzales/mynotesapp/Note.kt
@@ -1,11 +1,11 @@
 package com.ggonzales.mynotesapp
 
 class Note{
-    var noteID : Int? = null
+    var noteID : Long? = null
     var noteTitle : String? = null
     var noteDesc : String? = null
 
-    constructor(noteID : Int, noteTitle: String, noteDesc : String){
+    constructor(noteID : Long, noteTitle: String, noteDesc : String){
         this.noteID = noteID
         this.noteTitle = noteTitle
         this.noteDesc = noteDesc
